requirements.txt
cymysql==0.9.2
django-cymysql==0.5.1
pkg-resources==0.0.0
pytz==2017.3
sqlparse==0.1.19
